feat(sentry): rename setup -> content-scripts
