+++ src/utils/sentry/content-script.ts
