+++ src/utils/sentry/popup.ts
@@ -4,7 +4,7 @@ import {
   getDefaultTags,
   client as sentryClient,
   scope as sentryScope
-} from "./setup";
+} from "./content-script";
 
 type ErrorBoundaryState = {
   hasError: boolean;
